HTTP client: Support retrying requests on non-fatal server errors
This also introduces more future-proof 'flags' parameters,
more capable convenience functions OSSL_HTTP_get_ex() and OSSL_HTTP_transfer_ex(),
the high-level function OSSL_HTTP_set_long_overall_timeout(), and
the low-level functions OSSL_HTTP_get_status() and OSSL_HTTP_may_retry().
The 'keep_alive' params are changed to 'flags' in a backward compatible way.

Fixes the handling of the 'server' argument and the default 'port' argument,
moving a check from OSSL_HTTP_set1_request() to OSSL_HTTP_open().

Generalize the 'server' argument of OSSL_HTTP_open() such that it may be a URL.
So the port may also be given this way, as a fallback to the 'port' argument.
This also fixes a minor bug:
A ':' that is part of an IPv6 host address was misinterpreted as a port separator.

Adds http_wget_test.c, which tests HTTP(S) GET with a real server if reachable.